Move Game Save settings to a new tab
