+++ web_src/sidepanel/EditorTab.jsx
@@ -3,29 +3,13 @@ import { Settings } from '../settings.js';
 import { Checkbox } from './Checkbox.jsx';
 import { setBuildMode, exportBuildChanges } from '../devBuildMode.js';
 
-export const DevBuildSettingsTab = () => {
+export const EditorTab = () => {
   // const [isDevModeEnabled, setIsDevMode] = useState(Settings.global.devMode);
   const [isBuildModeEnabled, setIsBuildMode] = useState(Settings.global.buildMode);
 
   return (
     <>
-      <h2>Developer settings</h2>
-
-      {/*
-       * Commented out for now, since Developer Mode doesn't actually do anything (yet)
-      <div style={{paddingBottom: "3em"}}>
-        <Checkbox
-          label="Developer Mode"
-          value="1"
-          checked={isDevModeEnabled}
-          onChange={() => {
-            setIsDevMode(!isDevModeEnabled);
-            setDevMode(!isDevModeEnabled);
-          }}
-        />
-      </div>
-      */}
-
+      <h2>Editing tools</h2>
       <div style={{ paddingBottom: '3em' }}>
         <Checkbox
           label="Build Mode"
